Integrated Move and Jump
Integrated Move and Jump with the new Input System using Ellen as the mesh for the main avatar.

NOTE Scivolo changed a bit in between Chris wrote the Canvas and the current version on the Unity Asset Store; this commit contains the (minimal) required changes to make everything work again.