Now Left CONTROL triggers Jumping
To avoid the ghosting of certain key combinations like W and SHIFT and SPACEBAR, we moved the LEFT CONTROL trigger from the Shooting action to the Jumping action.

NOTE: We left SPACEBAR to trigger Jumping as well, just in case a player has a better keyboard that doesn't ghost keys.